requirements.txt
pymodbus>=3.5,<4
pyserial>=3.5,<4
